+++ src/lib/libcmd/chgrp.c
@@ -147,7 +147,7 @@ getids(char* s, char** e, Key_t* key, int options)
 	key->uid = key->gid = -1;
 	while (isspace(*s))
 		s++;
-	for (t = s; (n = *t) && n != ':' && n != '.' && !isspace(n); t++);
+	for (t = s; (n = *t) && n != ':' && n != '.' && !isspace((int)n); t++);
 	if (n)
 	{
 		options |= OPT_CHOWN;
@@ -172,7 +172,7 @@ getids(char* s, char** e, Key_t* key, int options)
 			}
 			key->uid = j;
 		}
-		for (s = t; (n = *t) && !isspace(n); t++);
+		for (s = t; (n = *t) && !isspace((int)n); t++);
 		if (n)
 		{
 			if ((n = t++ - s) >= (ssize_t)sizeof(buf))
